add comments to front-end components
