+++ .gitignore
@@ -2,3 +2,4 @@
 *.todo.md
 node_modules/
 .next/
+yarn-error.log
