+++ package.json
@@ -22,6 +22,7 @@
     "react-dom": "^16.12.0"
   },
   "devDependencies": {
-    "prettier": "^1.19.1"
+    "prettier": "^1.19.1",
+    "prop-types": "^15.7.2"
   }
 }
