+++ src/components/ThreadsPage.js
@@ -1,62 +1,21 @@
 import fetch from 'isomorphic-unfetch'
-import Link from 'next/link'
 import { ROOT_URL } from '../constants'
 import { useState } from 'react'
 import { GlobalStyles } from './GlobalStyles'
 import { AddThreadForm } from './AddThreadForm'
+import { ThreadPreview } from './ThreadPreview'
 
-const Badge = ({ children, style }) => {
-  return (
-    <p
-      style={{
-        background: 'var(--neutral-050)',
-        display: 'inline',
-        padding: '4px',
-        ...style,
-      }}
-    >
-      {children}
-    </p>
-  )
-}
-
-const Thread = ({ id, title, description, comments, reactions }) => {
-  const reactionsCount = Object.values(reactions).reduce(
-    (sum, val) => val + sum,
-    0
-  )
-
-  const commentsCount = comments.length
-
-  return (
-    <div
-      style={{
-        border: '2px solid var(--neutral-100)',
-        marginTop: '8px',
-        padding: '16px',
-      }}
-    >
-      <h2 style={{ fontSize: '1.5rem', marginBottom: '12px' }}>
-        <Link href="/[threadId]" as={`/${id}`}>
-          <a>{title}</a>
-        </Link>
-      </h2>
-      <p>{description}</p>
-
-      <div style={{ display: 'flex' }}>
-        <Badge>💬 {commentsCount}</Badge>
-        <Badge style={{ marginLeft: '4px' }}>👍 {reactionsCount}</Badge>
-      </div>
-    </div>
-  )
-}
-
+/**
+ * Renders a list of all threads in the app as preview cards.
+ *
+ * Additionally, allows you to add a thread.
+ */
 export const ThreadsPage = ({ threads: initialThreads }) => {
   const [threads, setThreads] = useState(initialThreads)
   return (
     <div>
       {threads.map(thread => (
-        <Thread key={thread.id} {...thread} />
+        <ThreadPreview key={thread.id} {...thread} />
       ))}
 
       <div style={{ marginTop: '32px' }}>
@@ -79,6 +38,11 @@ export const ThreadsPage = ({ threads: initialThreads }) => {
   )
 }
 
+/**
+ * `getInitialProps` is NextJS's way of handling fetching data on the server.
+ *
+ * For more on `getInitialProps`, check out [this link](https://nextjs.org/docs/api-reference/data-fetching/getInitialProps)
+ */
 ThreadsPage.getInitialProps = async () => {
   const threads = await fetch(ROOT_URL + '/api/threads').then(res => res.json())
   return { threads }
